requirements.txt

# DEEP LEARNING & MACHINE LEARNING FRAMEWORKS
TensorFlow==2.13.0
numpy==1.24.3
scipy==1.11.1
scikit-learn==1.3.0

# IMAGE PROCESSING & COMPUTER VISION
opencv-python==4.8.0.74
Pillow==10.0.0
scikit-image==0.21.0
albumentations==1.3.1
imageio==2.31.1

# WEB FRAMEWORK & SERVER
Flask==2.3.2
Flask-CORS==4.0.0
Werkzeug==2.3.6
gunicorn==21.2.0
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.6

# DATA PROCESSING & ANALYSIS
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2

# UTILITIES & TOOLS
tqdm==4.65.0
python-dotenv==1.0.0
requests==2.31.0
